requirements.txt
mlflow==2.2.2
pandas
numpy
scikit-learn
matplotlib
seaborn
notebook
tqdm
types-PyYAML
PyYAML
fastapi
uvicorn

certifi